Move DataType and Facets to __base.py
