+++ src/pysdmx/model/__init__.py
@@ -15,6 +15,8 @@ from pysdmx.model.__base import (
     DataConsumer,
     DataflowRef,
     DataProvider,
+    DataType,
+    Facets,
     ItemReference,
     MetadataProvider,
     Organisation,
@@ -29,7 +31,7 @@ from pysdmx.model.code import (
     HierarchyAssociation,
     LevelType,
 )
-from pysdmx.model.concept import Concept, ConceptScheme, DataType, Facets
+from pysdmx.model.concept import Concept, ConceptScheme
 from pysdmx.model.constraint import (
     ConstraintAttachment,
     CubeKeyValue,
