+++ src/pysdmx/model/concept.py
@@ -10,152 +10,10 @@ can be defined using ``DataType``, ``Facets`` or enumeration (i.e. list of
 ``codes``).
 """
 
-from datetime import datetime
-from enum import Enum
-from typing import Iterator, Optional, Sequence, Union
+from typing import Iterator, Optional, Sequence
 
-from msgspec import Struct
-
-from pysdmx.model.__base import Item, ItemScheme
+from pysdmx.model.__base import DataType, Facets, Item, ItemScheme
 from pysdmx.model.code import Codelist
-
-
-class DataType(str, Enum):
-    """The expected data type for a concept."""
-
-    ALPHA = "Alpha"
-    """The alphabetic character set of A-Z, a-z.."""
-    ALPHA_NUM = "AlphaNumeric"
-    """The character set of A-Z, a-z and 0-9."""
-    BASIC_TIME_PERIOD = "BasicTimePeriod"
-    """The ISO standard periods."""
-    BIG_INTEGER = "BigInteger"
-    """Immutable arbitrary-precision signed integer."""
-    BOOLEAN = "Boolean"
-    """True or False."""
-    COUNT = "Count"
-    """A simple incrementing integer type."""
-    DATE = "GregorianDay"
-    """A ISO 8601 date (e.g. ``2011-06-17``)."""
-    DATE_TIME = "DateTime"
-    """An ISO 8601 point in time."""
-    DAY = "Day"
-    """A day in the ISO 8601 calendar (e.g. ``---31``)."""
-    DECIMAL = "Decimal"
-    """Immutable arbitrary-precision signed decimal number."""
-    DOUBLE = "Double"
-    """A decimal number (8 bytes)."""
-    DURATION = "Duration"
-    """An ISO 8601 duration."""
-    FLOAT = "Float"
-    """A decimal number (4 bytes)."""
-    GREGORIAN_TIME_PERIOD = "GregorianTimePeriod"
-    """This is the union of YEAR, YEAR_MONTH, and DATE."""
-    INCREMENTAL = "Incremental"
-    """An incrementable value (either a whole or decimal number)."""
-    INTEGER = "Integer"
-    """A whole number (4 bytes)."""
-    LONG = "Long"
-    """A whole number (8 bytes)."""
-    MONTH = "Month"
-    """A month in the ISO 8601 calendar (e.g. ``--12``)."""
-    MONTH_DAY = "MonthDay"
-    """A month day in the ISO 8601 calendar (e.g. ``--12-31``)."""
-    NUMERIC = "Numeric"
-    """The simple numeric character set of 0-9, treated as a string."""
-    PERIOD = "ObservationalTimePeriod"
-    """A reporting period. The format varies with the frequency."""
-    REP_DAY = "ReportingDay"
-    """The SDMX reporting period for daily frequency (e.g. 2025-D001)."""
-    REP_MONTH = "ReportingMonth"
-    """The SDMX reporting period for monthly frequency (e.g. 2025-M01)."""
-    REP_QUARTER = "ReportingQuarter"
-    """The SDMX reporting period for quarterly frequency (e.g. 2025-Q1)."""
-    REP_SEMESTER = "ReportingSemester"
-    """The SDMX reporting period for half-yearly frequency (e.g. 2025-S1)."""
-    REP_TRIMESTER = "ReportingTrimester"
-    """The SDMX reporting period for trimestrial frequency (e.g. 2025-T1)."""
-    REP_WEEK = "ReportingWeek"
-    """The SDMX reporting period for weekly frequency (e.g. 2025-W01)."""
-    REP_YEAR = "ReportingYear"
-    """The SDMX reporting period for yearly frequency (e.g. 2025-A1)."""
-    SHORT = "Short"
-    """A whole number (2 bytes)."""
-    STD_TIME_PERIOD = "StandardTimePeriod"
-    """The ISO standard periods and the SDMX reporting periods."""
-    STRING = "String"
-    """A string (as immutable sequence of Unicode code points)."""
-    TIME = "Time"
-    """An ISO 8601 time (e.g. ``12:50:42``)."""
-    URI = "URI"
-    """A uniform resource identifier, such as a URL."""
-    XHTML = "XHTML"
-    """XHTML (or HTML) markup."""
-    YEAR = "GregorianYear"
-    """An ISO 8601 year (e.g. ``2000``)."""
-    YEAR_MONTH = "GregorianYearMonth"
-    """An ISO 8601 year and month (e.g. ``2000-01``)."""
-
-    def __str__(self) -> str:
-        """Data Type String representation."""
-        return self.value
-
-    def __repr__(self) -> str:
-        """Data Type String representation."""
-        return f"{self.__class__.__name__}.{self._name_}"
-
-
-class Facets(Struct, frozen=True, omit_defaults=True, repr_omit_defaults=True):
-    """Additional information about the concept expected values.
-
-    The facets that apply vary with the type. For example,
-    ``min_mength`` and ``max_length`` are typically used with
-    strings, while ``start_value`` and ``end_value`` are typically
-    used with numeric sequences.
-    """
-
-    min_length: Optional[int] = None
-    """The minimum length for the value (e.g. ``1`` character)."""
-    max_length: Optional[int] = None
-    """The maximum length for the value (e.g. ``256`` characters)."""
-    min_value: Optional[Union[int, float]] = None
-    """Used by ranges to indicate what the lower bound of the range is."""
-    max_value: Optional[Union[int, float]] = None
-    """Used by ranges to indicate what the upper bound of the range is."""
-    start_value: Optional[Union[int, float]] = None
-    """Used in a numeric sequence to indicate its starting point."""
-    end_value: Optional[Union[int, float]] = None
-    """Used in a numeric sequence to indicate its ending point."""
-    interval: Optional[Union[int, float]] = None
-    """The permitted interval (increment) in a sequence."""
-    time_interval: Optional[str] = None
-    """The permitted duration in a time sequence."""
-    decimals: Optional[int] = None
-    """The number of characters allowed after the decimal separator."""
-    pattern: Optional[str] = None
-    """A regular expression the value must match."""
-    start_time: Optional[datetime] = None
-    """Indicates the starting point of a sequence."""
-    end_time: Optional[datetime] = None
-    """Indicates the ending point of a sequence."""
-    is_sequence: bool = False
-    """Whether the values are intended to be ordered."""
-    is_multilingual: bool = False
-    """Whether the text can be in multiple languages."""
-
-    def __str__(self) -> str:
-        """Custom string representation without the class name."""
-        processed_output = []
-        for attr, value, *_ in self.__rich_repr__():  # type: ignore[misc]
-            processed_output.append(f"{attr}: {value}")
-        return f"{', '.join(processed_output)}"
-
-    def __repr__(self) -> str:
-        """Custom __repr__ that omits empty sequences."""
-        attrs = []
-        for attr, value, *_ in self.__rich_repr__():  # type: ignore[misc]
-            attrs.append(f"{attr}={repr(value)}")
-        return f"{self.__class__.__name__}({', '.join(attrs)})"
 
 
 class Concept(Item, frozen=True, omit_defaults=True, tag=True):
